refactor(sources): split location importer in multiple files
LIIKUNTA-578.
The Location-importer source code file had become some long,
that it was really hard to read and understand.
Now the content is split into several files depending on the type of content.

The imports are changed from relative imports to absolutes from the changed scope.
