+++ sources/ingest/importers/tests/test_accessibility_shortcomings.py
@@ -1,7 +1,7 @@
 import pytest
 
-from ..location import (
-    AccessibilityShortcoming,
+from ingest.importers.location.dataclasses import AccessibilityShortcoming
+from ingest.importers.location.utils import (
     get_unit_id_to_accessibility_shortcomings_mapping,
 )
 
@@ -38,7 +38,7 @@ MOCKED_SERVICE_MAP_UNIT_VIEWPOINT_RESPONSE = {
 @pytest.fixture
 def mocked_service_map_unit_viewpoint_response(mocker):
     return mocker.patch(
-        "ingest.importers.location.request_json",
+        "ingest.importers.location.utils.request_json",
         return_value=MOCKED_SERVICE_MAP_UNIT_VIEWPOINT_RESPONSE,
     )
